add redirect on browser reload
