+++ src/components/simulation/FooterButtons.vue
@@ -2,17 +2,7 @@
 import { useSimulationStore } from '@/stores/simulations'
 import { useRouter } from 'vue-router'
 import { computed } from 'vue'
-import { useAddressStore } from '@/stores/address'
-import { useRoofsStore } from '@/stores/roof'
-import { useConsumptionAndProductionStore } from '@/stores/consumptionAndProduction'
-import { useAutocalsolStore } from '@/stores/autocalsol'
-import { useSolarPanelStore } from '@/stores/solarPanels'
-
-const addressStore = useAddressStore()
-const roofStore = useRoofsStore()
-const consumptionAndProductionStore = useConsumptionAndProductionStore()
-const autocalsolStore = useAutocalsolStore()
-const solarPanelStore = useSolarPanelStore()
+import { resetStoresAndRedirect } from '@/services/resetStores'
 
 const props = defineProps<{
   isDisplayNextButton: Boolean
@@ -27,12 +17,7 @@ function clickButtonCancel() {
       'Cette action vous renvoie en début de simulation, vos données actuelles seront effacées'
     )
   ) {
-    addressStore.resetAddress()
-    roofStore.resetRoofStore()
-    consumptionAndProductionStore.resetAnnualConsumption()
-    autocalsolStore.resetAutocalsolResult()
-    solarPanelStore.resetPanelStore()
-
+    resetStoresAndRedirect()
     router.push('/roof-selection')
   }
 }
