+++ src/router/index.ts
@@ -1,3 +1,4 @@
+// import { useRoofsStore } from '@/stores/roof'
 import { createRouter, createWebHistory } from 'vue-router'
 
 const routes = [
@@ -48,6 +49,30 @@ const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: routes,
 })
+
+// router.beforeEach((to, from, next) => {
+//   console.log('avant le routing')
+//   console.log('from', from.name)
+//   console.log('to', to.name)
+//   if (from.name === undefined && to.name !== 'home') {
+//     next({ name: 'home' })
+//   } else {
+//     next()
+//   }
+// })
+
+// const roofStore = useRoofsStore()
+
+// router.beforeEach((to, from, next) => {
+//   // if( roofStore.selectedBuildingId === null) {
+//     console.log ("store vide")
+//     // console.log(roofStore.selectedBuildingId)
+//     // next({ name: 'roof-selection' })
+//   // }
+//   // else {
+//     next()
+//   // }
+// })
 
 export { routes }
 export default router
